+++ tests/test_android_here.py
@@ -1,5 +1,4 @@
 import pytest
-from pathlib import Path
 
 
 @pytest.mark.parametrize(